[BE] refactor(#443): ranking refactor
* 랭킹 존재하지 않으면 0반환

+++ backend/src/main/java/com/example/backend/auth/api/service/rank/response/UserRankingResponse.java
@@ -11,5 +11,5 @@ public class UserRankingResponse {
 
     private int score;
 
-    private int ranking;
+    private Long ranking;
 }
